Enabled client side validation after server side.
This has been disabled when we are developing the server side
validation. We are providing the client and server side validation
because the user can disable the javascript validation from developer
option or from browser setting which will stop the validation of input
so to handle that we are also implanting server side validation.