Added Email and phone features
Added Email and phone features in signUp page and this fields are
present in communication class where phone is a another class which
contains country code and mobileNumber and we need to format this by
using format with implementing the format interface in a class and we
need to override the 2 methods and we need to format this incoming
string from signup page from this 91-987678986 to country code : 91 and
mobileNumber : 987678986 and we need to return the phone object.
To work this formatter we need to implement a interface inside a
configuration class which is WebMvcConfiguration and then we need to
override the default method name addFormatter and then we need to
registry.addFormatter(new phoneNumberFormatter()).