Created BillContoller with Amount and creditCard properties.
Created billController and added Amount and CreditCard property and we
have implemented the amount and credit card bean formating by which we
are formatting the required input in string format and formatting it to
beans and then converting it to the property bean and with implementing
the formatter interface to the property classes and overriding the print
and parse methods to show output and parsing input. And we have used
Locale classes for the inside this formatter class as it is overriding
the method with same fields in parameters and we are used this locale to
change the set of input like 10 USD to  $ 10.