Created own Age annotation with mandatory annotations
Created own Age annotation with mandatory annotations like @Retention
and Target and then we have created a Constraint class and added the
constraint class as annotation parameter and this constraint annotation
is used for the our custom business logic.

In next commit we will add the messages for this annotation will be
added to properties file and sprig will take those messages from there
after performing some beans in side configuration class.