Using session for accessing user details in every JSP page adding once
Here we have used session by which we are creating session key and value
which is a server side storing cookie by which we can also ad cookie
expire time and if we want till server get off we can do it.
here we are using this session as cookie in server side by just
providing a key to client by which it will be able to access the server
side data stored.