Noting changed - Difference between @SessionAttributes and HttpSession
When we want to use the session information which is stored inside the
server we can directly use it in JSP page and when we want to perform
any Business logic with that session value then we need to access
directly within same controller if we want to access in different
controller then we need to use @SessionAttribute("KEY") Object obj
inside the handermethod parameter. but if we have used SessionStatus
Object in main controller one of the handler method parameter and set to
completed i.e sessionStstus.setComplete() then the session attribute is
limited to that controller till that method only. Even after reloading
in same method of same controller it will not give value, it will give
empty string in JSP. If we want to store data permanently inside session
then Spring will recommend to use HttpSession instead SeesionAttribute
annotation.
We can access this HttpSession in all controller and if we want to
perform any business logic then we need to use session.getAttribute() by
providing HttpSession object inside handler method parameters. and we
can directly access in any JSP page with key by which we have stored
value inside httpSession object. This HttpSession is widely used as
compared to SessionAttribute as this SessionAttribute is limited because
we can only add Model object with same names inside SessionAttribute so
this is limited and we can add anything inside the Model object by using
setAttribute() method. And we can also remove the stored information
from HttpSession by using httpSession.invalidate().