Email Sending with spring boot stater mail dependency
Added Email sending with spring boot stater mail dependency with
javaMailSender interface and JavaMailSender class we have creating in
configuration class and then we are implementing email service clayer
with interface and class and here we are using MimeMessage class to
create mimemessage object with javamailSender object and then with the
help of MimeMessageHelper we creating message sender object with name
help to send mail with mimemessager object and true for multipart and to
use UTF-8 as 3 parameters.
then adding subject body text and email to and then we are sending email
with javaMailSenderImpl with send() method by passing mimeMessage object
adding all details in helper object of mimeMessageHelper reference and
send the mimeMessage object as parameter to send message of
javaMailSenderImpl.

This MimeMessage is coming from jakarta.mail.internet.MimeMessage and
handing exceptions with MessagingException.