Added a model attribute to check whether the user disabled cookies
Here we have added a dummy model attribute by which we are checking
whether the user browser is disabled to access cookies because we are
storing cookies with a unique number by which the spring will search the
number in server and get the stored cookies inside the server this
HttpSession or SessionAttribute will store the cookies inside the server
by storing only a cookie with a unique number and we are checking this
by accessing this session values by using getAttribute() method. and if
the value is null that means the server is unable to fetch the unique
cookie from user browser as cookies is disabled by user so we will show
a page for enabling the cookies.

Made changes in UserInfoDTO by adding a property as relationResult to
store result  which is coming from LCAppServiceImpl and then so we need
to use Model inside of @ModelAttribute because we are storing
information inside @SessionAttribute so we need to put DTO class after
@Valid and then we need to add attribute manually and we also need to
add the BindingResult value inside the attribute like this to get error
message when user do any mistake in JSP page
model.addAttribute(BindingResult.MODEL_KEY_PREFIX+"userInfo",
bindingResult);
Here userInfo is a DTO that we need to change as per the DTO we are
using for our JSP page.
And we have added the c: tag library inside the process page and used
this to use at link by using this c tags are used to parse the URL and
with this we can give only the jsp page uri without context and this
will carry the cookies with in url if and only if cookies are disabled
by user or else it will not carry the cookies with url.

In this we have added a relation class in which we are saving the
constant values of relation result and we have divided method in side
LCAppServiceImpl which will find the relation.