Feedback service is implemented
Feedback service is implemented which will send response to user saying
'Thank you for sending feedback' and feedback to developer saying
'Feedback from [user] with [response message]'. We are implementing
JavaMailSenderImpl class with all properties set inside configuration
class and using MimeMessage class object and MimeMessageHelper class by
setting properties as (mimeMessageObj, true(for MultiPart to set the
message with attachments and alternative text and inline text), UTF-8).
And by using mimeMessageHelper object we are setting properties like
subject, To, Text, and so on. and then we need to send message by using
javaMailSenderImpl object with send method and sending mimeMessagehelper
object as arguments.